Extract NumberFormat.format(string). Add same for MoneyFormat.format. Share string splitting logic with input transformation.

+++ app/composeApp/src/commonMain/kotlin/com/alexrdclement/palette/app/demo/formats/money/MoneyFormatDemo.kt
@@ -88,7 +88,7 @@ class MoneyFormatDemoState(
         internal set
 
     val text by derivedStateOf {
-        moneyFormat.format(demoTextFieldState.text.toString().toDouble())
+        moneyFormat.format(demoTextFieldState.text.toString())
     }
 }
